LayoutTests/media/audio-constructor-src-expected.txt

TEST(audio instanceof HTMLAudioElement) OK
EVENT(load) TEST(relativeURL(audio.currentSrc)=='content/test.wav') FAIL
EVENT(play)
END OF TEST
